+++ lib/src/lib/module/directive/is-loading.directive.module.ts
@@ -1,7 +1,7 @@
 import { NgModule } from '@angular/core';
 import { IsLoadingSpinnerComponent } from './is-loading-spinner.component';
 import { IsLoadingDirective } from './is-loading.directive';
-import { IsLoadingService } from '../is-loading.service';
+import { IsLoadingService } from '../../is-loading.service';
 
 @NgModule({
   declarations: [IsLoadingDirective, IsLoadingSpinnerComponent],